refactor: AuthenticationScheme 파일 enums 패키지 이동
